Rip Old Scripts
Rewrote our BitManager and BitGridManager Scripts a bit. No longer using coroutines, it would but out with high bit rates, now updating in udpate() with a delay to the animation speed (animTimer) might need an increase since 10,000 or 100,000 bitrate causes it to break. But yeh Bitrate refreshes only on an activation call to a method in BitManager (i think) one of our 2 scripts you should be able to find it. But yeh, works better than before i think so maybe improvement im still kinda skeptical if it works or not as usual.