Big Change (Actually tho)
So... lots to go through, not sure how best to explain, scripts we went through are in the changelog.
BitManager now "listens" for OnBitrateChanged to invoke the change in the GlobalBitRate and handles Stat Changes.
BasicUpgrade now add stats to the CoreStats script and updates the upgradeInfo to upgrade the info UI.
CoreStats were overhauled, checks for branch type and creates a dictionary "StatMap" of dynamically allocated variable names for different stats.
UpdateInfoPanel now updates the color of the info panel based on upgrade branch path
UpgradeInfo now updates the InfoPanel with new upgrade stats and effects (no level milestones or passiveeffecttext updating yet)
And yeah thats pretty much it, new UpgradeMainButtonHandler listens for the selected upgrade node and checks to make sure we can afford the upgrade.