UpgradeTrackerManager set up
So uhhh not very visually like dev friendly, but it should track and store all upgrade nodes levels, it grabs them at the start since its gonna grab most of them anyways, but it checks for unlocked levels and whatnot. Uhhh system sweep works, haven't confirmed further level milestones beyond lvl 5, but we need to implement some sort of NodeConnector script to know from which knows where we can go Up or Down, maybe if thats how we can set it up. but yeh should be good and we moved around the bit grids to be bottom right aligned for visual clarity away from the rest of the Ui components.