Lots A Progers
NodeConnector is Build. A bit overkill for our purposes but it gets the job done dynamically for all nodes and paths so far. Colors could be adjusted per usual but more towards polish. Only things missing are spawning BitGrids when unlocking them Lvl 1 (Cept Alpha) and uhhh Custom scripts for the rest of the upgrades :D good luck hermano