Music and UI Buttons
So MusicManager and MusicHelper (For Audio UI Panel) are working in tandem with one another surprisingly. Forgot to commit our changes from yesterday but we changed up our "meta" buttons to be a 4 quadrant system and carried those over to our titlescreen to diversify our buttons and whatnot. Music manager is pretty handy and visually helpful in the inspector. All Audio meta buttons are working, shuffling is a touch buggy with the PlayingObject indexing using the original playlist ordering and not the new shuffled order, but we could also maybe change our list visually to "shuffle" aswell to make the play objects line up better? Sliders still need implementations but might get to that when we maybe get to SFX and whatnot.